Clean up ECO-2404
