+++ src/typescript/frontend/src/lib/store/event/event-store.ts
@@ -11,49 +11,29 @@ import {
 } from "@sdk/indexer-v2/types";
 import { createStore } from "zustand/vanilla";
 import { immer } from "zustand/middleware/immer";
-import { type SetLatestBarsArgs, type EventStore } from "./types";
+import { type SetLatestBarsArgs, type EventStore } from "../../../../../types";
 import {
   ensureMarketInStore,
   handleLatestBarForPeriodicStateEvent,
   handleLatestBarForSwapEvent,
   pushPeriodicStateEvents,
   toMappedMarketEvents,
+  initialState,
 } from "./utils";
-import { initialStatePatch } from "./local-storage";
 import { periodEnumToRawDuration } from "@sdk/const";
 import { createWebSocketClientStore, type WebSocketClientStore } from "../websocket/store";
 import { DEBUG_ASSERT, extractFilter } from "@sdk/utils";
-import { parseJSON, stringifyJSON } from "utils";
-
-export const TIMEOUT = 1000 * 60;
-
-type EventLocalStorageKey = "swap" | "chat" | "periodic" | "market" | "liquidity";
-
-const updateLocalStorage = (key: EventLocalStorageKey, event: AnyEventModel) => {
-  const str = localStorage.getItem(key) ?? "[]";
-  const data: AnyEventModel[] = parseJSON(str);
-  data.unshift(event);
-  localStorage.setItem(key, stringifyJSON(data));
-};
-
-const cleanReadLocalStorage = (key: EventLocalStorageKey) => {
-  const str = localStorage.getItem(key) ?? "[]";
-  const data: AnyEventModel[] = parseJSON(str);
-  const relevants = data.filter(
-    (e) => e.transaction.timestamp > new Date(new Date().getTime() - TIMEOUT)
-  );
-  localStorage.setItem(key, stringifyJSON(relevants));
-  return relevants;
-};
-
-const clearLocalStorage = (key: EventLocalStorageKey) => {
-  localStorage.setItem(key, "[]");
-};
+import {
+  updateLocalStorage,
+  cleanReadLocalStorage,
+  clearLocalStorage,
+  LOCAL_STORAGE_EVENT_TYPES,
+} from "./local-storage";
 
 export const createEventStore = () => {
   const store = createStore<EventStore & WebSocketClientStore>()(
     immer((set, get) => ({
-      ...initialStatePatch(),
+      ...initialState(),
       getMarket: (emojis) => get().markets.get(emojis.join("")),
       getRegisteredMarkets: () => {
         return get().markets;
@@ -188,9 +168,8 @@ export const createEventStore = () => {
     }))
   );
 
-  const eventTypes: EventLocalStorageKey[] = ["swap", "chat", "liquidity", "market", "periodic"];
   const state = store.getState();
-  for (const eventType of eventTypes) {
+  for (const eventType of LOCAL_STORAGE_EVENT_TYPES) {
     try {
       const events = cleanReadLocalStorage(eventType);
       for (const event of events) {
